refactor(quiz): split code
