+++ functions/quiz.json
@@ -1,0 +1,1 @@
+[{"comment":" O CEO eh um cara muito bacana e tambem doidao. A empresa esta experimentando com um modelo de gestao totalmente horizontal ja faz um tempo e esta dando bons resultados. As pessoas propoem aumentos pros proprios salarios em uma reuniao aberta. Ambiente legal e com muita autonomia. O unico problema eh a localizacao do escritorio que fica na Vila Andrade","answer":"2BOmGEFG8HrZzaaslQz/i4k8Vgo=","alternatives":["ACCERA","EASY TAXI","CONCORRENTE EMPIRICUS","EDUK"]},{"comment":" Maior concentração de pessoas feias da região da Paulista.Vc só é valorizado se for formado nas faculdades de playboy (FGV, Mackenzie), pobre só tem vez pra ser burro de carga.","answer":"9dEqPFshgUsdQUPoTE2s27P1TZs=","alternatives":["COBLI","CONTAAZUL","CROSOFTEN","*"]},{"comment":" A empresa é bem interessante para se trabalhar e bastante inclusiva. Comportamentos relacionados a qqr preconceito ou racismo sao (realmente) severamente punidos. Ponto extremamente positivo, pois parece que ser babaca e nazista se tornou um comportamento padrão em várias startups e empresas de TI!Entretanto, a gerencia parece bem desconectada em relacao ao que acontece no tocante ao trabalho e dia a dia com os funcionários que estao alocados em clientes (e é pior ainda pros funcionários que ficam ou vao pros EUA), quase que literalmente \"tocando o foda-se\" e dizendo que o gerente do cliente tem poder total sobre a gente e eles nao podem fazer nada se eles decidem algo). A sede no Silicon Valley parece ser praticamente outra empresa completamente diferente dos escritorios no Brasil.Ainda sobre gerencia, fora dos clientes, os assistentes da gerencia e administração (os chamados buddies) se desdobram tentando auxiliar os subordinados, porem eles tem pouquissimo poder na empresa e nem sempre contam com assistencia. A empresa tem bons (e verdadeiros) beneficos, porem você e praticamente se torna um escravo do trabalho e nao dura muito na empresa, dependendo do cliente em que você cai quando é contratado. Se tiver sorte, você trabalha em um ótimo cliente e vive feliz para sempre. Os beneficios entao podem n compensar, depende mto de colocar td na balança... As politicas de RH da empresa sao quase zero, e os processos seletivos internos sao tocados de uma maneira extremamente ridicula e com funcionarios de RH extremamente despreparados, literalmente te desanimando e desmotivando de participar de qualquer um deles. Essa parte piorou exponencialmente desde a época que eu entrei na empresa.","answer":"7A2iJujQyfe6Ri0mLelBcYS1Kfk=","alternatives":["AVENUE CODE","QUINTO ANDAR","HOTMART ","CUOR DI CREMA"]},{"comment":" Desorganizada, com problemas sérios de gestão, RH péssimo, muita promessa e pouco comprometimento. Agora está ficando cada vez pior, após a venda para a Didi (chinesa).","answer":"Wk6Vk8PwCZhykT1x1/8IDH/UtyY=","alternatives":["NEOWAY","EBANX","MASTERTECH","99"]},{"comment":" Empresa extremamente machista, não contratam mulheres. Mandei um curriculum, e eles me chamaram para entrevista. Como o meu nome é Darci,  eles não conseguiram distinguir meu gênero. Na chegada, o youtuber Rodrigo ( que se acha o CEO, mas não passa de um youtuber) ficou surpreso pelo fato de eu ser uma mulher, e então me passou um teste muito dificil, que ninguém conseguiria passar, somente para justificar seu machismo e não me contratar. Não tem mulheres nesta empresa. ","answer":"J2LCZx1XJXglp4lfj0reS3qUf8c=","alternatives":["BLINKTRADE","CABIFY","NEON PAGAMENTOS S/A ","IFOOD"]},{"comment":" o que mais me assusta na empresa é que ela passou por uma aceleradora no mesmo tempo do guia bolso, pagar.me, creditas... e essas empresas hoje estão  anos luz a frente. E a ____ não consegue crescer.ceo não sabe lidar com as pessoas e não se preocupa com elas.  os livros de empreendedorismo parecem não ensinar isso. faz muito discurso sobre cultura, mas só da boca para fora. isso causa uma falta de confiança na figura dele.Um dos co-fundadores tem tato com as pessoas e acredito que ele seria um bom ceo e é o que mais apresenta ter os valores que estão escritos em uma parede.o trabalho no dia a dia é até bom devido as pessoas que estão ali. Tem possibilidade de home office.  falsa liberdade e autonomia, é apenas da boca para fora, igual a cultura. e as decisões demoram a ser tomadas.Os salários são abaixo do mercado (exceto para os desenvolvedores). ","answer":"bwFByjKzZquwhcYpOcoiCQoebb0=","alternatives":["PLING","CABIFY","AGENDOR","ROCK CONTENT"]},{"comment":" Empresa pequena, que no começo parece a Disney, tudo mágico e perfeito, mesa de sinuca, churrasco na sexta, mas basta alguns dias de trabalho pra ver que tudo é ilusão. Rotina da galera pesada, cobrança rotineira de hora extra, tudo muito individualista, projetos sem foco mudando constantemente conforme o humor do CEO, que mais parece uma criança com muito dinheiro fazendo o que bem entende sem se preocupar em ouvir a opinião de seus funcionários, acha que entende de design mas no fim a única coisa que fazemos é criar cópias do Google! ","answer":"hG/Lp7p9wsiOV4NXP5WxqadshHs=","alternatives":["CABIFY","TELIGGO (MYLUCCY)","PLING","CONQUER"]},{"comment":" Uma escola, onde as ideias podem mesmo ser colocadas em prática. Ambiente agradável, com pessoas do bem e com potencial ...a startup está virando uma empresa e está aprendendo a lidar com essa nova fase. Nem tudo é um mar de rosas, mas também não é de espinhos.  ","answer":"azyRMl4/Qdo824Swct866HX5XRA=","alternatives":["BRUNO & ANDRÉ MITOS","OLIST","RESULTADOS DIGITAIS","CLICKBUS"]},{"comment":" Empresa que cresceu rapidamente mas com pouca estrutura e preparo para seguir com o crescimento. A plataforma é uma colcha de retalhos, cada um faz o que quer. Quem não é do time de tecnologia, é remunerado bem a baixo da média. Rola bastante preconceito e quem não for amigo dos donos, acaba sendo escanteado.","answer":"XoCiOIA6gPLohUNwSduZBVv9PeI=","alternatives":["EBANX","DECORA","SINGU","STONE"]},{"comment":" Claramente é uma empresa para deixar o CEO rico. Eu odiava enganar o cliente vendendo um sonho e sabendo que aquilo nunca iria ser entregue. Sua VP de vendas era estagiaria até pouco tempo, acho que ela tem um  caso com alguém da diretoria. O CEO parece um avatar do Lazy Town. Odiei trabalhar lá, eu ganhava muito, muito mal mesmo.","answer":"XLMPdYtaL+mqPkE3sI+Hw57WCwY=","alternatives":["APPLE (SETOR APPSTORE)","FAZINOVA (ESCOLA DA BELPESCE)","PLING","OUTBOUND MARKETING REEV"]}]
